requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
matplotlib>=3.7.0
pillow>=9.4.0
numpy>=1.24.0
opencv-python>=4.7.0
argparse>=1.4.0

